api/requirements.txt
Python==3.14.0
Flask==3.1.2
Werkzeug==3.1.4
python-dotenv==1.0.1
psycopg2-binary==2.9.11
